backend/requirements.txt
aiosqlite==0.21.0
fastapi==0.115.12
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20
SQLAlchemy==2.0.41
uvicorn==0.34.2


facenet-pytorch==2.5.3
faiss-cpu==1.11.0
numpy==2.2.5
pillow==11.2.1
torch==2.7.0